作文批閱/requirements.txt
streamlit
pandas
numpy
plotly
yfinance
matplotlib
google-generativeai